backend/requirements.txt
django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
gunicorn
